HttpClient, HttpRequest, HttpResponse, Uri. +retrieve X-Transmission-Session-Id header.

+++ transmission_rpc.h
@@ -5,7 +5,7 @@
 
 struct TorrentProgress {
     std::string name;
-    float percentage;
+    double percentage;
 };
